Add cow_http_hd:parse_sec_websocket_version_client/1
From RFC6455.

The code is slower than a plain binary_to_integer(Version),
because it enforces that the value is in the 0..255 range as
per the specification, and that the input binary is no longer
than 3 characters.

It is however better to enforce this because the client could
also send a huge integer and we don't really want to convert it
and waste resources.