Add support for chunked transfer-encoding trailers
It considers all 0-sized chunks that aren't \r\n\r\n
to be trailers. There's no option for enabling/disabling
the behavior (for example when the te header was sent).
It doesn't parse the trailer, it's up to the user to
parse it separately via the new cow_http:headers/1 functions.

Note that this reuses the TotalLength part of the returned
'done' tuple to signal whether there are trailers. This value
has been ignored in Cowboy since 2.0 and was just a historical
leftover. I'm not aware of anyone using this module outside of
Gun or Cowboy, so I don't expect this to break anything. If it
does, well, it's not a documented function anyway. Your fault.