Add HTTP/2 option stream_window_data_threshold
This controls how large the window should be before we are
willing to send data. We of course always send data when
the data we need to send is lower than the current window.

This is both an optimization and a fix for the data dribble
denial of service vulnerability (CVE-2019-9511).