Optimize HPACK Huffman decoding
The previous naive implementation ended up not benefitting
from binary match context reuse because it was matching on
bit boundaries (and not bytes). This new implementation
matches a byte at a time.

The result is almost twice faster when decoding the examples
from the HPACK RFC (before/after):

cow_hpack:decode_huffman in 0.250666s
cow_hpack:decode_huffman in 0.133743s