Initial HTTP/3 and QPACK implementation
Cowlib now uses GitHub Actions for CI. As a result
of this change, Cowlib is tested against OTP-24+.

This commit adds initial implementations of
cow_http3, cow_http3_machine and cow_qpack.

Because QPACK is similar to HPACK, some encoding and
decoding functions were moved to a common include file,
particularly the huffman functions.

The cow_http module now contains the types and functions
common to all or most versions of HTTP. The types and
functions specific to HTTP/1 were moved to the new
cow_http1 module.

Because HTTP/3 is similar to HTTP/2, part of the code
processing headers is common and can be found in
cow_http. Other functions common to both versions
were moved out of cow_http2_machine.

This commit updates comments indicating that the HTTP/2
PRIORITY mechanism will no longer be implemented.