Optimise cow_uri:urldecode
The recently added `json` module in OTP demonstrated
new, better ways of optimising binary string manipulations.
Some of which likely make better use of recent compiler
or VM improvements.

The main improvement is returning as fast as possible
when there's nothing to decode:

  cow_uri:urldecode in 0.053833s
  cow_uri:urldecode in 0.005991s

When there's a little to decode, the performance is
nicely improved, although not as much:

  cow_uri:urldecode_hex in 0.061018s
  cow_uri:urldecode_hex in 0.040028s

When the input if fully/mostly encoded JP text, the
difference is remarkable:

  cow_uri:urldecode_jp_hex in 0.099333s
  cow_uri:urldecode_jp_hex in 0.039391s

  cow_uri:urldecode_jp_mixed_hex in 0.097653s
  cow_uri:urldecode_jp_mixed_hex in 0.046385s

The worst case scenario unfortunately is slower than
before, but nothing too bad. The worst case is when
the input is a sequence of 3 encoded characters
followed by 3 non-encoded characters:

  cow_uri:urldecode_worst_case_hex in 0.050017s
  cow_uri:urldecode_worst_case_hex in 0.061515s